% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/augFun.Est.R
\name{augFun.Est}
\alias{augFun.Est}
\title{Machine learning estimate for the augmented function.

'augFun.Est()' returns the augmented function a() at some given observations.

This function is used to estimate the augmented term in the augmentation estimation
function using a library of prediction algorithms.}
\usage{
augFun.Est(Y, Trt, Z, W, newZ, newW, Pi, SL.lib, SL.family)
}
\arguments{
\item{Y}{a vector for outcome}

\item{Trt}{a vector for treatment indicator}

\item{Z}{a vector for biomarker}

\item{W}{a matrix for baseline covariates (not inlcuing the biomarker)}

\item{newZ}{biomarker value to be evaluated}

\item{newW}{baseline covariate values to be evaluated}

\item{Pi}{probability of receiving the treatment (T=1)}

\item{SL.lib}{a list of functions for candidate prediction algorithms}

\item{SL.family}{allows gaussian or binomial to describe the error distribution}
}
\value{
A list with the first argument being the estimate of the interested
parameter beta based on the unaugmented estimation function and the second argument being
being the augmented term a() evaluated at the given observation (newZ, newW)

@export
}
\description{
Machine learning estimate for the augmented function.

'augFun.Est()' returns the augmented function a() at some given observations.

This function is used to estimate the augmented term in the augmentation estimation
function using a library of prediction algorithms.
}
