% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/augFunA.R
\name{augFunA}
\alias{augFunA}
\title{Estimation of the first agumented term in the two-phase design by indiivdual
machine learning prediction algorithms}
\usage{
augFunA(Y, Trt, Z, R, W, Pi, probR, newZ, newW, SL.lib, SL.family, b.fun)
}
\arguments{
\item{Y}{a vector for outcome}

\item{Trt}{a vector for treatment indicator}

\item{Z}{a vector for biomarker}

\item{R}{an indicator for whether the biomarker is observed}

\item{W}{a matrix for baseline covariates (not inlcuing the biomarker)}

\item{Pi}{probability of receiving the treatment (T=1)}

\item{probR}{conditional probability for the event that biomarker is observed}

\item{newZ}{a new vector for biomarker at which the agumented function a() is evaluated}

\item{newW}{a new matrix for baseline covariates at which the agumented function a() is evaluated}

\item{SL.lib}{a list of functions for candidate prediction algorithms}

\item{SL.family}{allows gaussian or binomial to describe the error distribution}

\item{b.fun}{the second augmented term which is a function of (outcome, treatment, baseline covairate)}
}
\value{
a list with the first argument being the initial estimate of beta without augmentation and
the second argument being the estimate of the augmented function a() at (newZ, newW)
}
\description{
'augFunA()' returns the estimate of first agumented function a() in the two-phase
design by indiivdual machine learning prediction algorithms
}
\details{
This function is used to estimate the first agumented term of the estimation function
for the two-phase design. This agumented function is a function of (biomarker, baseline covariate).
}
