% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/augFunB.R
\name{augFunB}
\alias{augFunB}
\title{Estimation of the second agumented term in the two-phase design by indiivdual
machine learning prediction algorithms}
\usage{
augFunB(
  Y,
  Trt,
  Z,
  R,
  W,
  Pi,
  probR,
  newY,
  newTrt,
  newW,
  SL.lib,
  SL.family,
  a.fun
)
}
\arguments{
\item{Y}{a vector for outcome}

\item{Trt}{a vector for treatment indicator}

\item{Z}{a vector for biomarker}

\item{R}{an indicator for whether the biomarker is observed}

\item{W}{a matrix for baseline covariates (not inlcuing the biomarker)}

\item{Pi}{probability of receiving the treatment (T=1)}

\item{probR}{conditional probability for the event that biomarker is observed}

\item{newY}{a new vector for outcome at which the agumented function b() is evaluated}

\item{newTrt}{a new vector for treatment indicator at which the agumented function b() is evaluated}

\item{newW}{a new matrix for baseline covariates at which the agumented function b() is evaluated}

\item{SL.lib}{a list of functions for candidate prediction algorithms}

\item{SL.family}{allows gaussian or binomial to describe the error distribution}

\item{a.fun}{the first augmented term which is a function of (biomarker, baseline covairate)}
}
\value{
a list with the first argument being the initial estimate of beta without augmentation and
the second argument being the estimate of the augmented function b() at (newY, newTrt, newW)
}
\description{
'augFunB()' returns the estimate of second agumented function b(Y, T, W) in the two-phase
design by indiivdual machine learning prediction algorithms
}
\details{
This function is used to estimate the second agumented term of the estimation function
for the two-phase design. This agumented function is a function of (outcome,treatment,
baseline covariate).
}
