% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BetaVar.R
\name{BetaVar}
\alias{BetaVar}
\title{Sandwich variance estimation for the biomarker studies in the two-phase setting.}
\usage{
BetaVar(Y, Trt, Z, R, Pi, probR, bet, a.fun, b.fun)
}
\arguments{
\item{Y}{a vector for outcome}

\item{Trt}{a vector for treatment indicator}

\item{Z}{a vector for biomarker}

\item{R}{an indicator for whether the biomarker is observed}

\item{Pi}{probability of receiving the treatment (T=1)}

\item{probR}{conditional probability for the event that biomarker is observed}

\item{bet}{parameter estimate}

\item{a.fun}{the first augmented term which is a function of (biomarker, baseline covairate)}

\item{b.fun}{the first augmented term which is a function of (outcome, treatment, baseline covairate)}
}
\value{
the variance estimate of the parameter beta
}
\description{
'BetaVar()' returns the variance estimate of the parameter beta in the two-phase setting.
}
\details{
This function is used to calculate the sandwich variance estimate for the parameter estimate
in the two-phase setting.
}
