% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calVar.R
\name{calVar}
\alias{calVar}
\title{Sandwich variance estimation for the biomarker studies in the complete-data setting.}
\usage{
calVar(Y, Trt, Z, Pi, bet, a.fun)
}
\arguments{
\item{Y}{a vector for outcome}

\item{Trt}{a vector for treatment indicator}

\item{Z}{a vector for biomarker}

\item{Pi}{probability of receiving the treatment (T=1)}

\item{bet}{parameter estimate}

\item{a.fun}{augmented function evaluated at the observation (Y, Trt, Z)}
}
\value{
the variance estimate of the parameter beta
}
\description{
'calVar()' returns the variance estimate of the parameter beta in the complete-data setting.
}
\details{
This function is used to calculate the sandwich variance estimate for the parameter estimate
in the complete-data setting.
}
