% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calWeight.R
\name{calWeight}
\alias{calWeight}
\title{Calculation of the weights for the super learner in the complete-data setting.}
\usage{
calWeight(Y, Trt, Z, W, Pi, SL.lib, SL.family, SL.method, SL.cv, ols.max)
}
\arguments{
\item{Y}{a vector for outcome}

\item{Trt}{a vector for treatment indicator}

\item{Z}{a vector for biomarker}

\item{W}{a matrix for baseline covariates (not inlcuing the biomarker)}

\item{Pi}{probability of receiving the treatment (T=1)}

\item{SL.lib}{a list of functions for candidate prediction algorithms}

\item{SL.family}{allows gaussian or binomial to describe the error distribution}

\item{SL.method}{a character value indicating the method used to estimate the coefficients
for the super learner and the model to combine the individual algorithms}

\item{SL.cv}{Number of splits for the cross-validation step in super learner}

\item{ols.max}{upper bound for the coefficients of individual algorithms in the super learner}
}
\value{
a vector of the coefficients for the individual algorithms in the library
}
\description{
'calWeight()' estimates the coefficients for the super learner and the model to combine
the individual algorithms in the library in the complete-data setting.
}
\details{
This function is used to estimate the weights for the super learner and the model to combine
the individual algorithms in the library in the complete-data setting.
}
