% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Der1.psi.R
\name{Der1.psi}
\alias{Der1.psi}
\title{First derivative of the estimating function without augmentation at a certain value.}
\usage{
Der1.psi(dat.vec, bet)
}
\description{
'Der1.psi()' is the first derivative value of the estimating function for the
individual observation at a certain value.
}
\details{
This function is used to calculate the first derivative of the estimating function
for the individual observation at a certain value.

@param dat.vec a vector contianing the observation of outcome, treatment, biomaker
@param bet value for the parameter beta

@return a value of the first derivative of the estimating function for the individual
observation (outcome, treatment, biomaker)

@example
dat.vec = c(1.5, 1, 0.3)
bet = 0.5
Der1.psi(dat.vec=dat.vec, bet=bet)

@export
}
