% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/invLogit.R
\name{invLogit}
\alias{invLogit}
\title{The inverse of the logistic function.}
\usage{
invLogit(u)
}
\arguments{
\item{u}{a real value}
}
\value{
a value in the interval \link{0,1}
}
\description{
'invLogit()' returns a value in the interval \link{0,1}.
}
\details{
This function is used to transform a value in the real line to the interval \link{0,1}.
}
\examples{
invLogit(2)

}
