% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/invLogitD1.R
\name{invLogitD1}
\alias{invLogitD1}
\title{The first derivative of the inverse of the logistic function.}
\usage{
invLogitD1(u)
}
\arguments{
\item{u}{a real value}
}
\value{
the first derivative value
}
\description{
'invLogitD1()' returns the first derivative value of the inverse of the logistic function.
}
\details{
This function is used to calculate the first derivative function for the inverse logit function.
}
\examples{
invLogitD1(2)

}
