% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psi.fun.R
\name{psi.fun}
\alias{psi.fun}
\title{Estimating function without augmentation for the parameters of interest.}
\usage{
psi.fun(dat.vec, bet)
}
\description{
'psi.fun()' is the value of the estimating function for the individual observation.
}
\details{
This function is used to calculate the estimating function for the individual observation.

@param dat.vec a vector contianing the observation of outcome, treatment, biomaker
@param bet value for the parameter beta

@return a value of the estimating function for the individual observation (outcome, treatment, biomaker)

@example
dat.vec = c(1.5, 1, 0.3)
bet = 0.5
psi.fun(dat.vec=dat.vec, bet=bet)

@export
}
