% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psiStar.R
\name{psiStar}
\alias{psiStar}
\title{Estimating function with augmentation for the parameters of interest in two-phase design.}
\usage{
psiStar(Y, Trt, Z, R, Pi, probR, bet, a.fun, b.fun)
}
\arguments{
\item{Y}{a vector for outcome}

\item{Trt}{a vector for treatment indicator}

\item{Z}{a vector for biomarker}

\item{R}{an indicator for whether the biomarker is observed}

\item{Pi}{probability of receiving the treatment (T=1)}

\item{probR}{conditional probability for the event that biomarker is observed}

\item{bet}{a vector containing the parameters of interest}

\item{a.fun}{the first augmented term which is a function of  (biomarker, baseline covairate)}

\item{b.fun}{the second augmented term which is a function of (outcome, treatment, baseline covairate)

@return a matrix containing the estimating function at each individual observation, in which
the row represents the subjects, the column represents the parameters

@export}
}
\description{
'psiStar()' is estimation function with augmentation in the two-phase design.
}
\details{
This function is used to calculate the estimation function with augmentation in the two-phase design.
}
