% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psiStarVar.R
\name{psiStarVar}
\alias{psiStarVar}
\title{Variance estimate for the estimating function with augmentation in the two-phase design.}
\usage{
psiStarVar(Y, Trt, Z, R, Pi, probR, bet, a.fun, b.fun)
}
\arguments{
\item{Y}{a vector for outcome}

\item{Trt}{a vector for treatment indicator}

\item{Z}{a vector for biomarker}

\item{R}{an indicator for whether the biomarker is observed}

\item{Pi}{probability of receiving the treatment (T=1)}

\item{probR}{conditional probability for the event that biomarker is observed}

\item{bet}{a vector containing the parameters of interest}

\item{a.fun}{the first augmented term which is a function of  (biomarker, baseline covairate)}

\item{b.fun}{the second augmented term which is a function of (outcome, treatment, baseline covairate)

@return a matrix for the variance estimate of the estimating function with augmentation
in the two-phase deisgn

@export}
}
\description{
'psiStarVar()' returns the variance estimate of the estimating function with augmentation in the two-phase design.
}
\details{
This function is used to calculate the variance of the augmented estimating function
in the two-phase design.
}
