% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SL2Phase.R
\name{SL2Phase}
\alias{SL2Phase}
\title{Efficient Estimation for biomarker studies in the two-phase sampling design.}
\usage{
SL2Phase(
  Y,
  Trt,
  Z,
  R,
  W,
  Pi,
  probR,
  SL.lib,
  SL.family,
  SL.method,
  ols.max,
  SL.cv,
  split.cv,
  iterMax,
  iterTol
)
}
\arguments{
\item{Y}{a vector for outcome}

\item{Trt}{a vector for treatment indicator}

\item{Z}{a vector for biomarker}

\item{R}{an indicator for whether the biomarker is observed}

\item{W}{a matrix for baseline covariates (not inlcuing the biomarker)}

\item{Pi}{probability of receiving the treatment (T=1)}

\item{probR}{conditional probability for the event that biomarker is observed}

\item{SL.lib}{a list of functions for candidate prediction algorithms}

\item{SL.family}{allows gaussian or binomial to describe the error distribution}

\item{SL.method}{a character value indicating the method used to estimate the coefficients
for the super learner and the model to combine the individual algorithms}

\item{ols.max}{upper bound for the coefficients of individual algorithms in the super learner}

\item{SL.cv}{Number of splits for the cross-validation step in super learner}

\item{iterMax}{maximum iteration numbers}

\item{iterTol}{iteration tolerance}

\item{Split.cv}{Number of splits for the calculation of cross-validated variance estimate}
}
\value{
A list with the first argument being the estimate of the association (beta) between the
outcome and biomarker, treatment indicator for the individual algorithms, as well as a super learner
that combines the individual algorithms, the second argument being the corresponding cross-validated
variance estimate.

@import SuperLearner rootSolve

@export
}
\description{
'SL2Phase()' returns efficient estimation  for biomarker studies in the two-phase
design in which the biomarker is only measured on a random sub-sample of subjects
chosen on the basis of (Y, T, W).

This function is used to estimate the parameter for the relationship between a
clinical outcome of interest and the biomarker, treatment indicator (can be omitted
in the case of a prognostic biomarker) in the two-phase setting. The efficient
estimation incorporates the independence between treatment and all relevant
baseline covariates.
}
