% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SLcomplete.R
\name{SLcomplete}
\alias{SLcomplete}
\title{Efficient estimation for biomarker studies embedded in randomized clinical trials
in the complete-data setting.

'SLcomplete()' returns efficient estimation for biomarker studies in the complete-
data setting in which the biomarker is measured for all subjects in the trial, by
using machine learning method.

This function is used to calculate estimates for the relationship
between a clinical outcome of interest and the biomarker, treatment indicator (can
be omitted in the case of a prognostic biomarker) in the complete-data setting, by
using the proposed efficient estimation approach. The efficient estimation incorporates
the independence between treatment and all relevant baseline covariates.}
\usage{
SLcomplete(
  Y,
  Trt,
  Z,
  W,
  Pi,
  SL.lib,
  SL.family,
  SL.method,
  SL.cv,
  Split.cv,
  ols.max
)
}
\arguments{
\item{Y}{a vector for outcome}

\item{Trt}{a vector for treatment indicator}

\item{Z}{a vector for biomarker}

\item{W}{a matrix for baseline covariates (not inlcuing the biomarker)}

\item{Pi}{probability of receiving the treatment (T=1)}

\item{SL.lib}{a list of functions for candidate prediction algorithms}

\item{SL.family}{allows gaussian or binomial to describe the error distribution}

\item{SL.method}{a character value indicating the method used to estimate the coefficients
for the super learner and the model to combine the individual algorithms}

\item{SL.cv}{Number of splits for the cross-validation step in super learner}

\item{Split.cv}{Number of splits for the calculation of cross-validated variance estimate}

\item{ols.max}{upper bound for the coefficients of individual algorithms in the super learner}
}
\value{
A list with the first argument being the estimate of the association (beta) between the
outcome and biomarker, treatment indicator for the individual algorithms, as well as a super learner
that combines the individual algorithms, the second argument being the corresponding cross-validated
variance estimate.

@import SuperLearner rootSolve

@export
}
\description{
Efficient estimation for biomarker studies embedded in randomized clinical trials
in the complete-data setting.

'SLcomplete()' returns efficient estimation for biomarker studies in the complete-
data setting in which the biomarker is measured for all subjects in the trial, by
using machine learning method.

This function is used to calculate estimates for the relationship
between a clinical outcome of interest and the biomarker, treatment indicator (can
be omitted in the case of a prognostic biomarker) in the complete-data setting, by
using the proposed efficient estimation approach. The efficient estimation incorporates
the independence between treatment and all relevant baseline covariates.
}
