% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SuperLearnerA.R
\name{SuperLearnerA}
\alias{SuperLearnerA}
\title{Machine learning estimation of the first agumented term in the two-phase design
using a super learner procedure}
\usage{
SuperLearnerA(
  Y,
  Trt,
  Z,
  R,
  W,
  Pi,
  probR,
  newZ,
  newW,
  SL.lib,
  SL.family,
  SL.method,
  b.fun,
  ols.max,
  SL.cv
)
}
\arguments{
\item{Y}{a vector for outcome}

\item{Trt}{a vector for treatment indicator}

\item{Z}{a vector for biomarker}

\item{R}{an indicator for whether the biomarker is observed}

\item{W}{a matrix for baseline covariates (not inlcuing the biomarker)}

\item{Pi}{probability of receiving the treatment (T=1)}

\item{probR}{conditional probability for the event that biomarker is observed}

\item{newZ}{a new vector for biomarker at which the agumented function a() is evaluated}

\item{newW}{a new matrix for baseline covariates at which the agumented function a() is evaluated}

\item{SL.lib}{a list of functions for candidate prediction algorithms}

\item{SL.family}{allows gaussian or binomial to describe the error distribution}

\item{SL.method}{a character value indicating the method used to estimate the coefficients
for the super learner and the model to combine the individual algorithms}

\item{b.fun}{the second augmented term which is a function of (outcome, treatment, baseline covairate)}

\item{ols.max}{upper bound for the coefficients of individual algorithms in the super learner}

\item{SL.cv}{Number of splits for the cross-validation step in super learner}
}
\value{
a list with the first argument being the estimate of augmented function a() by the individual
prediction algorithms and the second argument being the estimate by the super learner which combines
the individual algorithms.

@export
}
\description{
'SuperLearnerA()' returns the estimate of first agumented function a() in the two-phase
design using a super learner procedure
}
\details{
This function is used to estimate the first agumented term of the estimation function
for the two-phase design using a super learner procedure. This agumented function is
a function of (biomarker, baseline covariate).
}
