% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SuperLearnerB.R
\name{SuperLearnerB}
\alias{SuperLearnerB}
\title{Machine learning estimation of the second agumented term in the two-phase design
using a super learner procedure}
\usage{
SuperLearnerB(
  Y,
  Trt,
  Z,
  R,
  W,
  Pi,
  probR,
  newY,
  newTrt,
  newW,
  SL.lib,
  SL.family,
  SL.method,
  a.fun,
  ols.max,
  SL.cv
)
}
\arguments{
\item{Y}{a vector for outcome}

\item{Trt}{a vector for treatment indicator}

\item{Z}{a vector for biomarker}

\item{R}{an indicator for whether the biomarker is observed}

\item{W}{a matrix for baseline covariates (not inlcuing the biomarker)}

\item{Pi}{probability of receiving the treatment (T=1)}

\item{probR}{conditional probability for the event that biomarker is observed}

\item{newY}{a new vector for outcome at which the agumented function b() is evaluated}

\item{newTrt}{a new vector for treatment indicator at which the agumented function b() is evaluated}

\item{newW}{a new matrix for baseline covariates at which the agumented function b() is evaluated}

\item{SL.lib}{a list of functions for candidate prediction algorithms}

\item{SL.family}{allows gaussian or binomial to describe the error distribution}

\item{SL.method}{a character value indicating the method used to estimate the coefficients
for the super learner and the model to combine the individual algorithms}

\item{a.fun}{the first augmented term which is a function of (biomarker, baseline covairate)}

\item{ols.max}{upper bound for the coefficients of individual algorithms in the super learner}

\item{SL.cv}{Number of splits for the cross-validation step in super learner}
}
\value{
a list with the first argument being the estimate of augmented function b() by the individual
prediction algorithms and the second argument being the estimate by the super learner which combines
the individual algorithms.

@export
}
\description{
'SuperLearnerB()' returns the estimate of second agumented function b() in the two-phase
design using a super learner procedure
}
\details{
This function is used to estimate the second agumented term of the estimation function
for the two-phase design using a super learner procedure. This agumented function is
a function of (outcome, treatment, baseline covariate).
}
