% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Der1.psi.R
\name{Der1.psi}
\alias{Der1.psi}
\title{First derivative of the estimating function without augmentation at a certain value.}
\usage{
Der1.psi(dat.vec, bet)
}
\arguments{
\item{dat.vec}{a vector contianing the observation of outcome, treatment, biomaker}

\item{bet}{value for the parameter beta}
}
\value{
a value of the first derivative of the estimating function for the individual
observation (outcome, treatment, biomaker)
}
\description{
'Der1.psi()' is the first derivative value of the estimating function without
augmentation for the individual observation at a certain value.
}
\details{
This function is used to calculate the first derivative of the estimating function
without augmentation for the individual observation at a certain value.
}
