% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psi.fun.R
\name{psi.fun}
\alias{psi.fun}
\title{Estimating function without augmentation for the parameters of interest.}
\usage{
psi.fun(dat.vec, bet)
}
\arguments{
\item{dat.vec}{a vector contianing the observation of outcome, treatment, biomaker}

\item{bet}{value for the parameter beta}
}
\value{
a value of the estimating function for the individual observation (outcome, treatment, biomaker)
}
\description{
'psi.fun()' is the value of the estimating function for the individual observation.
}
\details{
This function is used to calculate the estimating function for the individual observation.
}
